+++ docs/CHANGES.rst
@@ -18,4 +18,5 @@ Main version of FeretUI, based on the POC
 * Added Response object
 * Added Session object
 * Added templating
+* Added static file register
 * Added translation behaviours